\name{adoptedLatentHeatOfVaporizationOfLiquidHe4}
\alias{adoptedLatentHeatOfVaporizationOfLiquidHe4}
\title{Adopted database for latent heat of vaporization of liquid 4He}
\usage{adoptedLatentHeatOfVaporizationOfLiquidHe4}
\description{\code{adoptedLatentHeatOfVaporizationOfLiquidHe4} is a table with the 
Adopted database for latent heat of vaporization of liquid He4
}
\source{
Donnelly R J
The Observed Properties of Liquid Helium at the Saturated Vapor Pressure
\url{http://pages.uoregon.edu/rjd/vapor17.htm}
}
\references{
Donnelly R J
The Observed Properties of Liquid Helium at the Saturated Vapor Pressure
\url{http://pages.uoregon.edu/rjd/vapor17.htm}
}
\author{Jose Gama}
\examples{
data(adoptedLatentHeatOfVaporizationOfLiquidHe4)
str(adoptedLatentHeatOfVaporizationOfLiquidHe4)
}
\keyword{datasets}
