\name{ AWGTOmm}
\alias{ AWGTOmm}
\title{Convert American wire gauge (SWG) to mm}
\usage{ AWGTOmm(n)}
\description{\code{ AWGTOmm} converts American wire gauge (SWG) to mm}
\arguments{
  \item{n}{ AWG gauge}
}
\value{
  \item{g}{gauge in mm}
}
\source{
rapidtables.com, 2014
convert American wire gauge (SWG) to mm
\url{http://www.rapidtables.com/calc/wire/awg-to-mm.htm}
}
\references{
rapidtables.com, 2014
convert American wire gauge (SWG) to mm
\url{http://www.rapidtables.com/calc/wire/awg-to-mm.htm}
}
\author{Jose Gama}
\keyword{programming}
