\name{coefficientsThermocoupleTypeB}
\alias{coefficientsThermocoupleTypeB}
\alias{coefficientsThermocoupleTypeE}
\alias{coefficientsThermocoupleTypeJ}
\alias{coefficientsThermocoupleTypeK}
\alias{coefficientsThermocoupleTypeN}
\alias{coefficientsThermocoupleTypeR}
\alias{coefficientsThermocoupleTypeS}
\alias{coefficientsThermocoupleTypeT}
\title{Polynomial Equation Coefficients for Voltage to Temperature for Thermocouple Type B}
\usage{coefficientsThermocoupleTypeB}
\description{\code{coefficientsThermocoupleTypeB} Coefficients for Voltage to Temperature for Thermocouple Type B
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(coefficientsThermocoupleTypeB)
str(coefficientsThermocoupleTypeB)
}
\keyword{datasets}
