\name{definitionThermocoupleTypes}
\alias{definitionThermocoupleTypes}
\title{Thermocouple Types Definitions}
\usage{definitionThermocoupleTypes}
\description{\code{definitionThermocoupleTypes} is a table with Thermocouple Types Definitions
}
\source{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_i.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_i.html}
}
\author{Jose Gama}
\examples{
data(definitionThermocoupleTypes)
str(definitionThermocoupleTypes)
}
\keyword{datasets}
