\name{ DiameterAWG}
\alias{ DiameterAWG}
\title{American Wire Gauge (AWG) diameter from AWG number}
\usage{ DiameterAWG(AWG)}
\description{\code{ DiameterAWG} Calculates American Wire Gauge (AWG) diameter from AWG number}
\arguments{
  \item{AWG}{AWG number}
}
\value{
  \item{d}{American Wire Gauge (AWG) diameter}
}
\source{
Lund Instrument Engineering, Inc., 2014
Wire Gauge and Current Limits Including Skin Depth and Strength
\url{http://www.powerstream.com/Wire_Size.htm}
}
\references{
Lund Instrument Engineering, Inc., 2014
Wire Gauge and Current Limits Including Skin Depth and Strength
\url{http://www.powerstream.com/Wire_Size.htm}
}
\author{Jose Gama}
\keyword{programming}
