\name{dielectricC.Density.ThermExpLiquid4HeSatVapPress}
\alias{dielectricC.Density.ThermExpLiquid4HeSatVapPress}
\title{Recommended values of the dielectric constant, density and thermal 
expansion coefficient of liquid 4He at saturated vapor pressure}
\usage{dielectricC.Density.ThermExpLiquid4HeSatVapPress}
\description{\code{dielectricC.Density.ThermExpLiquid4HeSatVapPress} is a table with the 
Recommended values of the dielectric constant, density and thermal expansion
 coefficient of liquid 4He at saturated vapor pressure
}
\source{
Donnelly R J
The Observed Properties of Liquid Helium at the Saturated Vapor Pressure
\url{http://pages.uoregon.edu/rjd/vapor2.htm}
}
\references{
Donnelly R J
The Observed Properties of Liquid Helium at the Saturated Vapor Pressure
\url{http://pages.uoregon.edu/rjd/vapor2.htm}
}
\author{Jose Gama}
\examples{
data(dielectricC.Density.ThermExpLiquid4HeSatVapPress)
str(dielectricC.Density.ThermExpLiquid4HeSatVapPress)
}
\keyword{datasets}
