\name{ds18B20TemperatureData}
\alias{ds18B20TemperatureData}
\title{Temperature/Data Relationship DS18B20}
\usage{ds18B20TemperatureData}
\description{\code{ds18B20TemperatureData} is a table with the Temperature/Data Relationship for the DS18B20
}
\source{
Maxim Integrated Products, Inc., 2014
DS18B20 datasheet REV: 042208
DS18B20 Programmable Resolution 1-Wire Digital Thermometer
}
\references{
Maxim Integrated Products, Inc., 2014
DS18B20 datasheet REV: 042208
DS18B20 Programmable Resolution 1-Wire Digital Thermometer
}
\author{Jose Gama}
\examples{
data(ds18B20TemperatureData)
str(ds18B20TemperatureData)
}
\keyword{datasets}
