\name{EquationThermocouplesTypeB}
\alias{EquationThermocouplesTypeB}
\alias{EquationThermocouplesTypeE}
\alias{EquationThermocouplesTypeJ}
\alias{EquationThermocouplesTypeK}
\alias{EquationThermocouplesTypeN}
\alias{EquationThermocouplesTypeR}
\alias{EquationThermocouplesTypeS}
\alias{EquationThermocouplesTypeT}
\title{Equation for Calculating Voltage from Temperature for Thermocouples Type B}
\usage{EquationThermocouplesTypeB(vT)}
\description{\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type B

\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type E

\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type J

\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type K

\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type N

\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type R

\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type S

\code{EquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type T
}
\arguments{
  \item{vT}{ Vector with temperatures (C)}
}
\value{
  \item{V}{Voltage (mV)}
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\keyword{programming}
