\name{ErrorLimitsThermocouples}
\alias{ErrorLimitsThermocouples}
\title{Limits of Error for Thermocouples}
\usage{ErrorLimitsThermocouples}
\description{\code{ErrorLimitsThermocouples} is a table with Limits of Error for Thermocouples}
\source{
Jim Strothman, 2006
ISA Handbook of Measurement Equations and Tables, 2nd Edition
The International Society of Automation
}
\references{
Jim Strothman, 2006
ISA Handbook of Measurement Equations and Tables, 2nd Edition
The International Society of Automation
}
\author{Jose Gama}
\examples{
data(ErrorLimitsThermocouples)
str(ErrorLimitsThermocouples)
}
\keyword{datasets}
