\name{fixedPointsITS90}
\alias{fixedPointsITS90}
\title{fixed Points ITS90}
\usage{fixedPointsITS90}
\description{\code{fixedPointsITS90} is a table with the fixed Points of ITS90
}
\source{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_iii.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_iii.html}
}
\author{Jose Gama}
\examples{
data(fixedPointsITS90)
str(fixedPointsITS90)
}
\keyword{datasets}
