\name{inverseCoefficientsThermocoupleTypeB}
\alias{inverseCoefficientsThermocoupleTypeB}
\alias{inverseCoefficientsThermocoupleTypeE}
\alias{inverseCoefficientsThermocoupleTypeJ}
\alias{inverseCoefficientsThermocoupleTypeK}
\alias{inverseCoefficientsThermocoupleTypeN}
\alias{inverseCoefficientsThermocoupleTypeR}
\alias{inverseCoefficientsThermocoupleTypeS}
\alias{inverseCoefficientsThermocoupleTypeT}
\title{Polynomial Equation Coefficients for Voltage to Temperature for Thermocouple Type B}
\usage{inverseCoefficientsThermocoupleTypeB}
\description{\code{inverseCoefficientsThermocoupleTypeB} Coefficients for Voltage to Temperature for Thermocouple Type B
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(inverseCoefficientsThermocoupleTypeB)
str(inverseCoefficientsThermocoupleTypeB)
}
\keyword{datasets}
