\name{InverseEquationThermocouplesTypeB}
\alias{InverseEquationThermocouplesTypeB}
\alias{InverseEquationThermocouplesTypeE}
\alias{InverseEquationThermocouplesTypeJ}
\alias{InverseEquationThermocouplesTypeK}
\alias{InverseEquationThermocouplesTypeN}
\alias{InverseEquationThermocouplesTypeR}
\alias{InverseEquationThermocouplesTypeS}
\alias{InverseEquationThermocouplesTypeT}
\title{Equation for Calculating Temperature from Voltage for Thermocouples Type B}
\usage{InverseEquationThermocouplesTypeB(vV)}
\description{\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type B

\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type E

\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type J

\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type K

\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type N

\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type R

\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type S

\code{InverseEquationThermocouplesTypeB} Calculates Voltage from Temperature for Thermocouples Type T
}
\arguments{
  \item{vV}{ Vector with voltages (C)}
}
\value{
  \item{T}{Temperature (C)}
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\keyword{programming}
