\name{inverseFunctionsThermocoupleRange}
\alias{inverseFunctionsThermocoupleRange}
\title{Error range for Polynomial inverse functions for Thermocouples}
\usage{inverseFunctionsThermocoupleRange}
\description{\code{inverseFunctionsThermocoupleRange} Error range for Polynomial inverse functions for Thermocouples
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(inverseFunctionsThermocoupleRange)
str(inverseFunctionsThermocoupleRange)
}
\keyword{datasets}
