\name{NominalSeebeckCoefficients}
\alias{NominalSeebeckCoefficients}
\title{Nominal Seebeck Coefficients}
\usage{NominalSeebeckCoefficients}
\description{\code{NominalSeebeckCoefficients} is a table with Nominal Seebeck Coefficients
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(NominalSeebeckCoefficients)
str(NominalSeebeckCoefficients)
}
\keyword{datasets}
