\name{PT100}
\alias{PT100}
\alias{PT1000}
\title{PT100 Resistance vs Temperature}
\usage{PT100}
\description{\code{PT100} is a table with PT100 resistance vs temperature

\code{PT1000} is a table with PT1000 resistance vs temperature
}
\source{
Pavitronic, 2014
pt100 resistance / temperature.
\url{http://pavitronic.dk/eng/pt100val.html}
}
\references{
Pavitronic, 2014
pt100 resistance / temperature.
\url{http://pavitronic.dk/eng/pt100val.html}
}
\author{Jose Gama}
\examples{
data(PT100)
str(PT100)
}
\keyword{datasets}
