\name{recommendedLatentHeatOfVaporizationOfLiquidHe4}
\alias{recommendedLatentHeatOfVaporizationOfLiquidHe4}
\title{Recommended values of the latent heat of vaporization of liquid 4He 
as a function of temperature at the saturated vapor pressure}
\usage{recommendedLatentHeatOfVaporizationOfLiquidHe4}
\description{\code{recommendedLatentHeatOfVaporizationOfLiquidHe4} is a table with the 
Recommended values of the latent heat of vaporization of liquid 4He as a function
 of temperature at the saturated vapor pressure
}
\source{
Donnelly R J
The Observed Properties of Liquid Helium at the Saturated Vapor Pressure
\url{http://pages.uoregon.edu/rjd/vapor17.htm}
}
\references{
Donnelly R J
The Observed Properties of Liquid Helium at the Saturated Vapor Pressure
\url{http://pages.uoregon.edu/rjd/vapor17.htm}
}
\author{Jose Gama}
\examples{
data(recommendedLatentHeatOfVaporizationOfLiquidHe4)
str(recommendedLatentHeatOfVaporizationOfLiquidHe4)
}
\keyword{datasets}
