\name{recommendedUpperTempLimitsProtectedThermocouples}
\alias{recommendedUpperTempLimitsProtectedThermocouples}
\title{Recommended Upper Temperature Limits for Protected Thermocouples}
\usage{recommendedUpperTempLimitsProtectedThermocouples}
\description{\code{recommendedUpperTempLimitsProtectedThermocouples} is a table with 
Recommended Upper Temperature Limits for Protected Thermocouples
}
\source{
Jim Strothman, 2006
ISA Handbook of Measurement Equations and Tables, 2nd Edition
The International Society of Automation
}
\references{
Jim Strothman, 2006
ISA Handbook of Measurement Equations and Tables, 2nd Edition
The International Society of Automation
}
\author{Jose Gama}
\examples{
data(recommendedUpperTempLimitsProtectedThermocouples)
str(recommendedUpperTempLimitsProtectedThermocouples)
}
\keyword{datasets}
