\name{ RTDalpha}
\alias{ RTDalpha}
\title{RTD alpha coefficient}
\usage{ RTDalpha(R0, R100)}
\description{\code{ RTDalpha} calculates RTD alpha coefficient}
\arguments{
  \item{R0}{resistance at 0C}
  \item{R100}{resistance at 100C}
}
\value{
  \item{alpha}{RTD alpha coefficient}
}
\source{
Capgo Inc., 2014
Introduction to RTDs
\url{http://www.capgo.com/Resources/Temperature/RTDs/RTD.html}
}
\references{
Capgo Inc., 2014
Introduction to RTDs
\url{http://www.capgo.com/Resources/Temperature/RTDs/RTD.html}
}
\author{Jose Gama}
\keyword{programming}
