\name{rtdAndThermistorStandardAccuracy}
\alias{rtdAndThermistorStandardAccuracy}
\title{Standard Accuracy for Thermocouples}
\usage{rtdAndThermistorStandardAccuracy}
\description{\code{rtdAndThermistorStandardAccuracy} is a table with standard accuracy for thermocouples
}
\source{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\references{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\author{Jose Gama}
\examples{
data(rtdAndThermistorStandardAccuracy)
str(rtdAndThermistorStandardAccuracy)
}
\keyword{datasets}
