\name{rtdAndThermistorStandardValues}
\alias{rtdAndThermistorStandardValues}
\title{Standard Values for Thermocouples}
\usage{rtdAndThermistorStandardValues}
\description{\code{rtdAndThermistorStandardValues} is a table with standard values for thermocouples
}
\source{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\references{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\author{Jose Gama}
\examples{
data(rtdAndThermistorStandardValues)
str(rtdAndThermistorStandardValues)
}
\keyword{datasets}
