\name{ RTDbeta}
\alias{ RTDbeta}
\title{RTD beta coefficient}
\usage{ RTDbeta(R0, Rtl, Tl, alpha, delta)}
\description{\code{ RTDbeta} calculates RTD beta coefficient}
\arguments{
  \item{R0}{resistance at 0C}
  \item{Rtl}{resistance of the sensor at the lowest temperature}
  \item{Tl}{lowest temperature in the calibration range}
  \item{alpha}{RTD alpha coefficient}
  \item{delta}{RTD delta coefficient}
}
\value{
  \item{beta}{RTD beta coefficient}
}
\source{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\references{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\author{Jose Gama}
\keyword{programming}
