\name{ RTDcoefficientA}
\alias{ RTDcoefficientA}
\alias{ RTDcoefficientB}
\alias{ RTDcoefficientC}
\title{RTD A coefficient}
\usage{ RTDcoefficientA(alpha, delta)}
\description{\code{ RTDcoefficientA} calculates RTD A coefficient

\code{ RTDcoefficientB} calculates RTD B coefficient

\code{ RTDcoefficientC} calculates RTD C coefficient}
\arguments{
  \item{alpha}{RTD alpha coefficient}
  \item{delta}{RTD delta coefficient}
}
\value{
  \item{A}{RTD A coefficient}
}
\source{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\references{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\author{Jose Gama}
\keyword{programming}
