\name{ RTDdelta}
\alias{ RTDdelta}
\title{RTD delta coefficient}
\usage{ RTDdelta(R0, Rth, Th,alpha)}
\description{\code{ RTDdelta} calculates RTD delta coefficient}
\arguments{
  \item{R0}{resistance at 0C}
  \item{Rth}{resistance of the sensor at the highest temperature}
  \item{Th}{highest temperature in the calibration range}
  \item{alpha}{RTD alpha coefficient}
}
\value{
  \item{delta}{RTD delta coefficient}
}
\source{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\references{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\author{Jose Gama}
\keyword{programming}
