\name{ RTDequation}
\alias{ RTDequation}
\title{RTD equation with 3 constants}
\usage{ RTDequation(R0, T, A, B, C=NA)}
\description{\code{ RTDequation} calculates the RTD equation}
\arguments{
  \item{R0}{ resistance at 0C}
  \item{T}{ temperature in C}
  \item{A}{ RTD constant}
  \item{B}{ RTD constant}
  \item{C}{ RTD constant}
}
\value{
  \item{R}{resistance}
}
\source{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\references{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\author{Jose Gama}
\keyword{programming}
