\name{RTDnickelIronResistanceFromAlpha}
\alias{RTDnickelIronResistanceFromAlpha}
\title{Simplified Equation for Nickel-Iron RTD Resistance}
\usage{RTDnickelIronResistanceFromAlpha(R0, T, alpha=NA)}
\description{\code{ RTDnickelIronResistanceFromAlpha} calculates simplified equation for Nickel-Iron RTD resistance}
\arguments{
  \item{R0}{resistance at 0C}
  \item{T}{ temperature in C}
  \item{alpha}{optional resistance's temperature coefficient}
}
\value{
  \item{R}{RTD resistance}
}
\source{
Capgo Inc., 2014
Introduction to RTDs
\url{http://www.capgo.com/Resources/Temperature/RTDs/RTD.html}
}
\references{
Capgo Inc., 2014
Introduction to RTDs
\url{http://www.capgo.com/Resources/Temperature/RTDs/RTD.html}
}
\author{Jose Gama}
\keyword{programming}
