\name{rtdPlatinumToleranceValues}
\alias{rtdPlatinumToleranceValues}
\title{Platinum RTD Tolerance Values}
\usage{rtdPlatinumToleranceValues}
\description{\code{rtdPlatinumToleranceValues} is a table with Platinum RTD Tolerance Values
}
\source{
Watlow Electric Manufacturing Company, 2014
Platinum RTD Tolerance Values
\url{https://www.watlow.com/reference/refdata/0315.cfm}
}
\references{
Watlow Electric Manufacturing Company, 2014
Platinum RTD Tolerance Values
\url{https://www.watlow.com/reference/refdata/0315.cfm}
}
\author{Jose Gama}
\examples{
data(rtdPlatinumToleranceValues)
str(rtdPlatinumToleranceValues)
}
\keyword{datasets}
