\name{rtdPT100}
\alias{rtdPT100}
\alias{rtdPT1000}
\title{rtdPT100 Resistance vs Temperature}
\usage{rtdPT100}
\description{\code{rtdPT100} is a table with PT100 resistance vs temperature

\code{rtdPT1000} is a table with PT1000 resistance vs temperature
}
\source{
Pavitronic, 2014
pt100 resistance / temperature.
\url{http://pavitronic.dk/eng/pt100val.html}
}
\references{
Pavitronic, 2014
pt100 resistance / temperature.
\url{http://pavitronic.dk/eng/pt100val.html}
}
\author{Jose Gama}
\examples{
data(rtdPT100)
str(rtdPT100)
}
\keyword{datasets}
