\name{rtdResistanceWireComparison}
\alias{rtdResistanceWireComparison}
\title{RTD Resistance Wire Comparison}
\usage{rtdResistanceWireComparison}
\description{\code{rtdResistanceWireComparison} is a table with RTD Resistance Wire Comparison
}
\source{
Watlow Electric Manufacturing Company, 2014
RTD Resistance Wire Comparison
\url{https://www.watlow.com/reference/refdata/0315.cfm}
}
\references{
Watlow Electric Manufacturing Company, 2014
RTD Resistance Wire Comparison
\url{https://www.watlow.com/reference/refdata/0315.cfm}
}
\author{Jose Gama}
\examples{
data(rtdResistanceWireComparison)
str(rtdResistanceWireComparison)
}
\keyword{datasets}
