\name{rtdResistivityAlpha}
\alias{rtdResistivityAlpha}
\title{Resistivity and Alpha Coefficients for RTDs}
\usage{rtdResistivityAlpha}
\description{\code{rtdResistivityAlpha} is a table with Resistivity and Alpha Coefficients for RTDs
}
\source{
Madur Inc., 2014
Resistive temperature detectors PTxx
\url{www.madur.com}
}
\references{
Madur Inc., 2014
Resistive temperature detectors PTxx
\url{www.madur.com}
}
\author{Jose Gama}
\examples{
data(rtdResistivityAlpha)
str(rtdResistivityAlpha)
}
\keyword{datasets}
