\name{ RTDtemperatureFromResistance}
\alias{ RTDtemperatureFromResistance}
\title{RTD Temperature from Resistance}
\usage{ RTDtemperatureFromResistance(R, R0)}
\description{\code{ RTDtemperatureFromResistance} calculates RTD Temperature from Resistance}
\arguments{
  \item{R}{resistance measured}
  \item{R0}{resistance at 0C}
}
\value{
  \item{T}{Temperature}
}
\source{
Madur In., 2014
Resistive temperature detectors PTxx
\url{www.madur.com}
}
\references{
Madur In., 2014
Resistive temperature detectors PTxx
\url{www.madur.com}
}
\author{Jose Gama}
\keyword{programming}
