\name{rtdTypes}
\alias{rtdTypes}
\title{Types of RTDs}
\usage{rtdTypes}
\description{\code{rtdTypes} is a table with Types of RTDs}
\source{
Capgo Inc., 2014
Introduction to RTDs
\url{http://www.capgo.com/Resources/Temperature/RTDs/RTD.html}
}
\references{
Capgo Inc., 2014
Introduction to RTDs
\url{http://www.capgo.com/Resources/Temperature/RTDs/RTD.html}
}
\author{Jose Gama}
\examples{
data(rtdTypes)
str(rtdTypes)
}
\keyword{datasets}
