\name{ SelfHeatingError}
\alias{ SelfHeatingError}
\title{self-heating error}
\usage{ SelfHeatingError(I, R, Ek)}
\description{\code{ SelfHeatingError} calculates the self-heating error}
\arguments{
  \item{I}{ intensity (A)}
  \item{R}{ resistance (ohm)}
  \item{Ek}{ self-heating coefficient(mW/C)}
}
\value{
  \item{E}{self-heating error}
}
\source{
Gerd Scheller, 2003
Error Analysis of a Temperature Measurement System 
with worked examples
JUMO, FAS 625, Edition 06.03
}
\references{
Gerd Scheller, 2003
Error Analysis of a Temperature Measurement System 
with worked examples
JUMO, FAS 625, Edition 06.03
}
\author{Jose Gama}
\keyword{programming}
