\name{ SensorSensitivity}
\alias{ SensorSensitivity}
\title{Sensitivity of the sensor}
\usage{ SensorSensitivity(T1, E1, T2, E2)}
\description{\code{ SensorSensitivity} calculates the Sensitivity of the sensor}
\arguments{
  \item{T1}{ measured temperature}
  \item{E1}{ resistance (platinum sensor) or the thermoelectric emf (thermocouple) for T1}
  \item{T2}{ measured temperature}
  \item{E2}{ resistance (platinum sensor) or the thermoelectric emf (thermocouple) for T2}
}
\value{
  \item{Cs}{Sensor Sensitivity}
}
\source{
Gerd Scheller, 2003
Error Analysis of a Temperature Measurement System 
with worked examples
JUMO, FAS 625, Edition 06.03
}
\references{
Gerd Scheller, 2003
Error Analysis of a Temperature Measurement System 
with worked examples
JUMO, FAS 625, Edition 06.03
}
\author{Jose Gama}
\keyword{programming}
