\name{singleLegThermoelements}
\alias{singleLegThermoelements}
\title{Letter designations, compositions, and trade names of single-leg thermoelements}
\usage{singleLegThermoelements}
\description{\code{singleLegThermoelements} is a table with the Letter designations, 
compositions, and trade names of single-leg thermoelements}
\source{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_ii.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_ii.html}
}
\author{Jose Gama}
\examples{
data(singleLegThermoelements)
str(singleLegThermoelements)
}
\keyword{datasets}
