\name{standardRTDandThermistorAccuracy}
\alias{standardRTDandThermistorAccuracy}
\title{Standard Accuracy for Thermocouples}
\usage{standardRTDandThermistorAccuracy}
\description{\code{standardRTDandThermistorAccuracy} is a table with standard accuracy for thermocouples
}
\source{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\references{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\author{Jose Gama}
\examples{
data(standardRTDandThermistorAccuracy)
str(standardRTDandThermistorAccuracy)
}
\keyword{datasets}
