\name{standardRTDandThermistorValues}
\alias{standardRTDandThermistorValues}
\title{Standard Values for Thermocouples}
\usage{standardRTDandThermistorValues}
\description{\code{standardRTDandThermistorValues} is a table with standard values for thermocouples
}
\source{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\references{
Veris Industries, 2009
Resources, Build-A-Sensor, Temperature Sensors
\url{http://www.veris.com/}
}
\author{Jose Gama}
\examples{
data(standardRTDandThermistorValues)
str(standardRTDandThermistorValues)
}
\keyword{datasets}
