\name{tableAWGCuWire}
\alias{tableAWGCuWire}
\title{AWG Wire Sizes with Resistance and More}
\usage{tableAWGCuWire}
\description{\code{tableAWGCuWire} is a table with AWG Wire sizes with resistance and more}
\source{
Lund Instrument Engineering, Inc., 2014
Wire Gauge and Current Limits Including Skin Depth and Strength
\url{http://www.powerstream.com/Wire_Size.htm}
}
\references{
Lund Instrument Engineering, Inc., 2014
Wire Gauge and Current Limits Including Skin Depth and Strength
\url{http://www.powerstream.com/Wire_Size.htm}
}
\author{Jose Gama}
\examples{
data(tableAWGCuWire)
str(tableAWGCuWire)
}
\keyword{datasets}
