\name{TemperatureDataDS18B20}
\alias{TemperatureDataDS18B20}
\title{Temperature/Data Relationship DS18B20}
\usage{TemperatureDataDS18B20}
\description{\code{TemperatureDataDS18B20} is a table with the Temperature/Data Relationship for the DS18B20
}
\source{
Maxim Integrated Products, Inc., 2014
DS18B20 datasheet REV: 042208
DS18B20 Programmable Resolution 1-Wire Digital Thermometer
}
\references{
Maxim Integrated Products, Inc., 2014
DS18B20 datasheet REV: 042208
DS18B20 Programmable Resolution 1-Wire Digital Thermometer
}
\author{Jose Gama}
\examples{
data(TemperatureDataDS18B20)
str(TemperatureDataDS18B20)
}
\keyword{datasets}
