\name{temperatureMeasurementDifficulty}
\alias{temperatureMeasurementDifficulty}
\title{temperature Measurement Difficulty}
\usage{temperatureMeasurementDifficulty}
\description{\code{temperatureMeasurementDifficulty} is a table with the current
state of difficulties with temperature measurements
}
\source{
CapGo, 2013
Is temperature measurement difficult?
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\references{
CapGo, 2013
Is temperature measurement difficult?
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\author{Jose Gama}
\examples{
data(temperatureMeasurementDifficulty)
str(temperatureMeasurementDifficulty)
}
\keyword{datasets}
