\name{temperatureSensorTypes}
\alias{temperatureSensorTypes}
\title{Temperature Sensor Types}
\usage{temperatureSensorTypes}
\description{\code{temperatureSensorTypes} is a table with Temperature Sensor Types
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(temperatureSensorTypes)
str(temperatureSensorTypes)
}
\keyword{datasets}
