\name{ ThermistorCalculateBeta}
\alias{ ThermistorCalculateBeta}
\title{Estimate thermistor beta coefficient from two known resistance/temperature values}
\usage{ ThermistorCalculateBeta(R0, T0, R1, T1)}
\description{\code{ ThermistorCalculateBeta} Estimates thermistor beta coefficient from two known resistance/temperature values
}
\arguments{
  \item{R0}{ resistance measurement 1}
  \item{T0}{ temperature measurement 1}
  \item{R1}{ resistance measurement 2}
  \item{T1}{ temperature measurement 2}
}
\value{
  \item{b}{beta coefficient}
}
\source{
RepRap wiki, 2014
Measuring Thermistor Beta
\url{http://reprap.org/wiki/MeasuringThermistorBeta}
}
\references{
RepRap wiki, 2014
Measuring Thermistor Beta
\url{http://reprap.org/wiki/MeasuringThermistorBeta}
}
\author{Jose Gama}
\keyword{programming}
