\name{ ThermistorCalibrationEquation}
\alias{ ThermistorCalibrationEquation}
\title{Thermistor calibration equation }
\usage{ ThermistorCalibrationEquation(R, R0, thCoeffs)}
\description{\code{ ThermistorCalibrationEquation} Thermistor calibration equation}
\arguments{
  \item{R}{ resistance measurement for temperature T}
  \item{R0}{ resistance measurement for temperature T0}
  \item{thCoeffs}{ Thermistor coefficient}
}
\value{
  \item{T}{temperature}
}
\source{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\references{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\author{Jose Gama}
\keyword{programming}
