\name{ ThermistorCalibrationEquationHoge3}
\alias{ ThermistorCalibrationEquationHoge3}
\title{Resistance–temperature calibration equation Hoge 1}
\usage{ ThermistorCalibrationEquationHoge3(Rt, A0, A1, A2, A3, A4)}
\description{\code{ ThermistorCalibrationEquationHoge3} Resistance–temperature 
calibration equation Hoge 3}
\arguments{
  \item{Rt}{ resistance measurement for temperature T}
  \item{A0}{ equation coefficient A0}
  \item{A1}{ equation coefficient A1}
  \item{A2}{ equation coefficient A2}
  \item{A3}{ equation coefficient A3}
  \item{A4}{ equation coefficient A4}
}
\value{
  \item{T}{temperature}
}
\source{
Chiachung Chen, 2009
Evaluation of resistance–temperature calibration equations for NTC thermistors
Measurement 42, Elsevier
}
\references{
Chiachung Chen, 2009
Evaluation of resistance–temperature calibration equations for NTC thermistors
Measurement 42, Elsevier
}
\author{Jose Gama}
\keyword{programming}
