\name{ ThermistorCalibrationEquationHoge4}
\alias{ ThermistorCalibrationEquationHoge4}
\title{Resistance–temperature calibration equation Hoge 1}
\usage{ ThermistorCalibrationEquationHoge4(Rt, A0, A1, A2, A5)}
\description{\code{ ThermistorCalibrationEquationHoge4} Resistance–temperature 
calibration equation Hoge 4}
\arguments{
  \item{Rt}{ resistance measurement for temperature T}
  \item{A0}{ equation coefficient A0}
  \item{A1}{ equation coefficient A1}
  \item{A2}{ equation coefficient A2}
  \item{A5}{ equation coefficient A5}
}
\value{
  \item{T}{temperature}
}
\source{
Chiachung Chen, 2009
Evaluation of resistance–temperature calibration equations for NTC thermistors
Measurement 42, Elsevier
}
\references{
Chiachung Chen, 2009
Evaluation of resistance–temperature calibration equations for NTC thermistors
Measurement 42, Elsevier
}
\author{Jose Gama}
\keyword{programming}
