\name{ ThermistorCalibrationEquationHoge5}
\alias{ ThermistorCalibrationEquationHoge5}
\title{Resistance–temperature calibration equation Hoge 1}
\usage{ ThermistorCalibrationEquationHoge5(Rt, C1, C2, C3)}
\description{\code{ ThermistorCalibrationEquationHoge5} Resistance–temperature 
calibration equation Hoge 1}
\arguments{
  \item{Rt}{ resistance measurement for temperature T}
  \item{C1}{ equation coefficient C1}
  \item{C2}{ equation coefficient C2}
  \item{C3}{ equation coefficient C3}
}
\value{
  \item{T}{temperature}
}
\source{
Chiachung Chen, 2009
Evaluation of resistance–temperature calibration equations for NTC thermistors
Measurement 42, Elsevier
}
\references{
Chiachung Chen, 2009
Evaluation of resistance–temperature calibration equations for NTC thermistors
Measurement 42, Elsevier
}
\author{Jose Gama}
\keyword{programming}
