\name{ ThermistorHoge1CoeffFromMeasurements}
\alias{ ThermistorHoge1CoeffFromMeasurements}
\alias{ ThermistorHoge2CoeffFromMeasurements}
\alias{ ThermistorHoge3CoeffFromMeasurements}
\alias{ ThermistorHoge4CoeffFromMeasurements}
\alias{ ThermistorAnyNumberOfCoeffFromMeasurements}
\title{Calculate Hoge1 coefficients from measurements}
\usage{ ThermistorHoge1CoeffFromMeasurements(resAndTemp)}
\description{\code{ ThermistorHoge1CoeffFromMeasurements} Calculate Hoge1 coefficients 
from measurements}
\arguments{
  \item{resAndTemp}{matrix with temperatures (C) in column 1 and 
resistance (ohm) in column 2}
}
\value{
  \item{b}{beta coefficient}
}
\author{Jose Gama}
\keyword{programming}
