\name{thermistorMaximumMeasuringVoltage}
\alias{thermistorMaximumMeasuringVoltage}
\title{NTC thermistor Nominal Resistance and Maximum measuring voltage}
\usage{thermistorMaximumMeasuringVoltage}
\description{\code{thermistorMaximumMeasuringVoltage} is a table with NTC thermistor 
Nominal Resistance (Rn) and Maximum measuring voltage (V)
}
\source{
AVX Corporation, 2014
AVX NTC Thermistors v11.4
\url{http://www.avx.com}
}
\references{
AVX Corporation, 2014
AVX NTC Thermistors v11.4
\url{http://www.avx.com}
}
\author{Jose Gama}
\examples{
data(thermistorMaximumMeasuringVoltage)
str(thermistorMaximumMeasuringVoltage)
}
\keyword{datasets}
