\name{ ThermistorResistanceSteinhartHart2}
\alias{ ThermistorResistanceSteinhartHart2}
\title{Steinhart-Hart equation for thermistor resistance, calculated with Maxima}
\usage{ ThermistorResistanceSteinhartHart2(T, A, B, C)}
\description{\code{ ThermistorResistanceSteinhartHart2} Steinhart-Hart equation for 
thermistor resistance, calculated with Maxima
}
\arguments{
  \item{T}{measured temperature for resistance R}
  \item{A}{Steinhart-Hart Coefficient A (K^0)}
  \item{B}{Steinhart-Hart Coefficient B (K^1)}
  \item{C}{Steinhart-Hart Coefficient C (K^2)}
}
\value{
  \item{R}{resistance}
}
\author{Jose Gama}
\keyword{programming}
