\name{ThermistorResistanceTolerance}
\alias{ThermistorResistanceTolerance}
\title{Thermistor relationship resistance tolerance}
\usage{ThermistorResistanceTolerance(TempAccy, alpha)}
\description{\code{ThermistorResistanceTolerance} Thermistor relationship resistance
 tolerance}
\arguments{
  \item{TempAccy}{Temperature Accuracy}
  \item{alpha}{Thermistor alpha constant}
}
\value{
  \item{t}{Tolerance}
}
\source{
Spectrum Sensors & Controls Inc., 2014
NTC Thermistors Engineering Notes
\url{http://www.SpecSensors.com}
}
\references{
Spectrum Sensors & Controls Inc., 2014
NTC Thermistors Engineering Notes
\url{http://www.SpecSensors.com}
}
\author{Jose Gama}
\keyword{programming}
