\name{ ThermistorSensitivity}
\alias{ ThermistorSensitivity}
\title{Thermistor Sensitivity}
\usage{ ThermistorSensitivity(T, beta)}
\description{\code{ ThermistorSensitivity} Thermistor Sensitivity
(relative change in resistance for a change in temperature)}
\arguments{
  \item{T}{measured temperature for resistance R}
  \item{beta}{beta Coefficient}
}
\value{
  \item{S}{Sensitivity}
}
\source{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\references{
John G. Webster and Halit Eren, 2014
Measurement, Instrumentation, and Sensors Handbook, Second Edition
Spatial, Mechanical, Thermal, and Radiation Measurement
CRC Press
}
\author{Jose Gama}
\keyword{programming}
