\name{ ThermistorSlope}
\alias{ ThermistorSlope}
\title{Thermistor Slope (Resistance Ratio)}
\usage{ ThermistorSlope(R0, R70)}
\description{\code{ ThermistorSlope} Thermistor Slope (Resistance Ratio)}
\arguments{
  \item{R0}{ resistance at temperature To (0C, expressed in Kelvin)}
  \item{R70}{ resistance at temperature To (70C, expressed in Kelvin)}
}
\value{
  \item{R}{resistance in ohms}
}
\source{
NTC Thermistor theory
BetaTHERM sensors
\url{www.betatherm.com}
}
\references{
NTC Thermistor theory
BetaTHERM sensors
\url{www.betatherm.com}
}
\author{Jose Gama}
\keyword{programming}
