\name{ ThermistorSteinhartHartCoeffFromMeasurements}
\alias{ ThermistorSteinhartHartCoeffFromMeasurements}
\title{Steinhart-Hart coefficients A, B, C from measurements}
\usage{ ThermistorSteinhartHartCoeffFromMeasurements(resAndTemp)}
\description{\code{ ThermistorSteinhartHartCoeffFromMeasurements} Steinhart-Hart 
coefficients A, B, C from measurements}
\arguments{
  \item{resAndTemp}{ matrix with temperatures (C) in column 1 and resistance (ohm) in column 2}
}
\value{
  \item{A}{ coefficient A}
  \item{B}{ coefficient B}
  \item{C}{ coefficient C}
}
\source{
NTC Thermistor theory
BetaTHERM sensors
\url{www.betatherm.com}
}
\references{
NTC Thermistor theory
BetaTHERM sensors
\url{www.betatherm.com}
}
\author{Jose Gama}
\keyword{programming}
