\name{ThermistorTemperatureAccuracy}
\alias{ThermistorTemperatureAccuracy}
\title{Thermistor relationship temperature accuracy}
\usage{ThermistorTemperatureAccuracy(ResTol, alpha)}
\description{\code{ThermistorTemperatureAccuracy} Thermistor relationship 
temperature accuracy}
\arguments{
  \item{ResTol}{ Thermistor resistance tolerance}
  \item{alpha}{ Thermistor alpha constant}
}
\value{
  \item{a}{Accuracy}
}
\source{
Spectrum Sensors & Controls Inc., 2014
NTC Thermistors Engineering Notes
\url{http://www.SpecSensors.com}
}
\references{
Spectrum Sensors & Controls Inc., 2014
NTC Thermistors Engineering Notes
\url{http://www.SpecSensors.com}
}
\author{Jose Gama}
\keyword{programming}
