\name{ ThermistorVolumeResistivityFromR25}
\alias{ ThermistorVolumeResistivityFromR25}
\title{Thermistor Volume Resistivity at 25C}
\usage{ ThermistorVolumeResistivityFromR25(R25, Thck, L, W)}
\description{\code{ ThermistorVolumeResistivityFromR25} Estimates thermistor Volume Resistivity at 25C
}
\arguments{
  \item{R25}{measured resistance 25C (ohms)}
  \item{Thck}{thickness of the conductor (chip) (cm)}
  \item{L}{length of the conductor (chip) (cm)}
  \item{W}{width of the conductor (chip) (cm)}
}
\value{
  \item{r}{Resistivity}
}
\source{
BetaTHERM sensors, 2014
NTC Thermistor theory
\url{www.betatherm.com}
}
\references{
BetaTHERM sensors, 2014
NTC Thermistor theory
\url{www.betatherm.com}
}
\author{Jose Gama}
\keyword{programming}
