\name{ ThermistorVolumeResistivityFromRho}
\alias{ ThermistorVolumeResistivityFromRho}
\title{Thermistor Volume Resistivity at 25C}
\usage{ ThermistorVolumeResistivityFromRho(Rho, Thck, L, W)}
\description{\code{ ThermistorVolumeResistivityFromRho} Estimates thermistor Volume Resistivity at 25C
}
\arguments{
  \item{Rho}{material resistivity in ohm/cm}
  \item{Thck}{thickness of the conductor (chip) (cm)}
  \item{L}{length of the conductor (chip) (cm)}
  \item{W}{width of the conductor (chip) (cm)}
}
\value{
  \item{r}{Resistivity}
}
\source{
BetaTHERM sensors, 2014
NTC Thermistor theory
\url{www.betatherm.com}
}
\references{
BetaTHERM sensors, 2014
NTC Thermistor theory
\url{www.betatherm.com}
}
\author{Jose Gama}
\keyword{programming}
