\name{ thermocoupleCables}
\alias{ thermocoupleCables}
\title{Thermocouple Cables}
\usage{ thermocoupleCables}
\description{\code{ thermocoupleCables} is a table with Thermocouple Cables}
\source{
Labfacility Limited, 2014
Thermocouple Cables
\url{https://www.labfacility.com/thermocouple-cables/}
}
\references{
Labfacility Limited, 2014
Thermocouple Cables
\url{https://www.labfacility.com/thermocouple-cables/}
}
\author{Jose Gama}
\keyword{datasets}
