\name{thermocoupleCoefficientsTypeB}
\alias{thermocoupleCoefficientsTypeB}
\alias{thermocoupleCoefficientsTypeE}
\alias{thermocoupleCoefficientsTypeJ}
\alias{thermocoupleCoefficientsTypeK}
\alias{thermocoupleCoefficientsTypeN}
\alias{thermocoupleCoefficientsTypeR}
\alias{thermocoupleCoefficientsTypeS}
\alias{thermocoupleCoefficientsTypeT}
\title{Polynomial Equation Coefficients for Voltage to Temperature for Thermocouple Type B}
\usage{thermocoupleCoefficientsTypeB}
\description{\code{thermocoupleCoefficientsTypeB} Coefficients for Voltage to Temperature for Thermocouple Type B
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(thermocoupleCoefficientsTypeB)
str(thermocoupleCoefficientsTypeB)
}
\keyword{datasets}
