\name{thermocoupleCoefficientsTypeBrationalPolynomial}
\alias{thermocoupleCoefficientsTypeBrationalPolynomial}
\alias{thermocoupleCoefficientsTypeJrationalPolynomial}
\alias{thermocoupleCoefficientsTypeKrationalPolynomial}
\alias{thermocoupleCoefficientsTypeErationalPolynomial}
\alias{thermocoupleCoefficientsTypeNrationalPolynomial}
\alias{thermocoupleCoefficientsTypeRrationalPolynomial}
\alias{thermocoupleCoefficientsTypeSrationalPolynomial}
\alias{thermocoupleCoefficientsTypeTrationalPolynomial}
\title{Polynomial Equation Coefficients for Voltage to Temperature for Thermocouple Type B}
\usage{thermocoupleCoefficientsTypeBrationalPolynomial}
\description{\code{thermocoupleCoefficientsTypeBrationalPolynomial} Polynomial Equation Coefficients 
for Voltage to Temperature for Thermocouple Type B
}
\source{
Mosaic Industries, Inc., 2014
rational polynomial function approximation for Type K thermocouples
\url{http://www.mosaic-industries.com/embedded-systems/microcontroller-projects/temperature-measurement/thermocouple/calibration-table#computing}-cold-junction-voltages
}
\references{
Mosaic Industries, Inc., 2014
rational polynomial function approximation for Type K thermocouples
\url{http://www.mosaic-industries.com/embedded-systems/microcontroller-projects/temperature-measurement/thermocouple/calibration-table#computing}-cold-junction-voltages
}
\author{Jose Gama}
\examples{
data(thermocoupleCoefficientsTypeBrationalPolynomial)
str(thermocoupleCoefficientsTypeBrationalPolynomial)
}
\keyword{datasets}
