\name{ thermocoupleColdJunctionVoltageCoeff}
\alias{ thermocoupleColdJunctionVoltageCoeff}
\title{Thermocouple Cold Junction Voltage Coefficients}
\usage{ thermocoupleColdJunctionVoltageCoeff}
\description{\code{ thermocoupleColdJunctionVoltageCoeff} is a table with Thermocouple Cold Junction Voltage Coefficients}
\source{
Capgo Pty Ltd, 2013
Computing cold junction voltages
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\references{
Capgo Pty Ltd, 2013
Computing cold junction voltages
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\author{Jose Gama}
\keyword{datasets}
