\name{thermocoupleDefinitionTypes}
\alias{thermocoupleDefinitionTypes}
\title{Thermocouple Types Definitions}
\usage{thermocoupleDefinitionTypes}
\description{\code{thermocoupleDefinitionTypes} is a table with Thermocouple Types Definitions
}
\source{
CapGo, 2013
Types of thermocouples
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\references{
CapGo, 2013
Types of thermocouples
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\author{Jose Gama}
\examples{
data(thermocoupleDefinitionTypes)
str(thermocoupleDefinitionTypes)
}
\keyword{datasets}
