\name{ThermocoupleEquationTypeB}
\alias{ThermocoupleEquationTypeB}
\alias{ThermocoupleEquationTypeE}
\alias{ThermocoupleEquationTypeJ}
\alias{ThermocoupleEquationTypeK}
\alias{ThermocoupleEquationTypeN}
\alias{ThermocoupleEquationTypeR}
\alias{ThermocoupleEquationTypeS}
\alias{ThermocoupleEquationTypeT}
\title{Equation for Calculating Voltage from Temperature for Thermocouples Type B}
\usage{ThermocoupleEquationTypeB(vT)}
\description{\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type B

\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type E

\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type J

\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type K

\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type N

\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type R

\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type S

\code{ThermocoupleEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type T
}
\arguments{
  \item{vT}{ Vector with temperatures (C)}
}
\value{
  \item{V}{Voltage (mV)}
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\keyword{programming}
