\name{thermocoupleErrorLimits}
\alias{thermocoupleErrorLimits}
\title{Limits of Error for Thermocouples}
\usage{thermocoupleErrorLimits}
\description{\code{thermocoupleErrorLimits} is a table with Limits of Error for Thermocouples}
\source{
Jim Strothman, 2006
ISA Handbook of Measurement Equations and Tables, 2nd Edition
The International Society of Automation
}
\references{
Jim Strothman, 2006
ISA Handbook of Measurement Equations and Tables, 2nd Edition
The International Society of Automation
}
\author{Jose Gama}
\examples{
data(thermocoupleErrorLimits)
str(thermocoupleErrorLimits)
}
\keyword{datasets}
