\name{ thermocoupleExtensionCables}
\alias{ thermocoupleExtensionCables}
\title{Thermocouple Extension Cables}
\usage{ thermocoupleExtensionCables}
\description{\code{ thermocoupleExtensionCables} is a table with Thermocouple Extension Cables}
\source{
Mike Nager, 2014
Designing with Thermocouples: Get the Most from Your Measurements
\url{www.phoenixcontact.com}
}
\references{
Mike Nager, 2014
Designing with Thermocouples: Get the Most from Your Measurements
\url{www.phoenixcontact.com}
}
\author{Jose Gama}
\keyword{datasets}
