\name{thermocoupleFixedPointsITS90}
\alias{thermocoupleFixedPointsITS90}
\title{fixed Points ITS90}
\usage{thermocoupleFixedPointsITS90}
\description{\code{thermocoupleFixedPointsITS90} is a table with the fixed Points of ITS90
}
\source{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_iii.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_iii.html}
}
\author{Jose Gama}
\examples{
data(thermocoupleFixedPointsITS90)
str(thermocoupleFixedPointsITS90)
}
\keyword{datasets}
