\name{thermocoupleInsulatingMaterialsCeramicPackedStock}
\alias{thermocoupleInsulatingMaterialsCeramicPackedStock}
\title{Insulating Materials for Ceramic Packed Thermocouple Stock}
\usage{thermocoupleInsulatingMaterialsCeramicPackedStock}
\description{\code{thermocoupleInsulatingMaterialsCeramicPackedStock} is a table with 
Insulating Materials for Ceramic Packed Thermocouple Stock}
\source{
American Society for Testing and Materials, 1981
Manual on the Use of Thermocouples in Temperature Measurement
Committee E20 on Temperature Measurement and Subcommittee E20.04 on Thermocouples
}
\references{
American Society for Testing and Materials, 1981
Manual on the Use of Thermocouples in Temperature Measurement
Committee E20 on Temperature Measurement and Subcommittee E20.04 on Thermocouples
}
\author{Jose Gama}
\examples{
data(thermocoupleInsulatingMaterialsCeramicPackedStock)
str(thermocoupleInsulatingMaterialsCeramicPackedStock)
}
\keyword{datasets}
