\name{thermocoupleInverseCoefficientsTypeB}
\alias{thermocoupleInverseCoefficientsTypeB}
\alias{thermocoupleInverseCoefficientsTypeE}
\alias{thermocoupleInverseCoefficientsTypeJ}
\alias{thermocoupleInverseCoefficientsTypeK}
\alias{thermocoupleInverseCoefficientsTypeN}
\alias{thermocoupleInverseCoefficientsTypeR}
\alias{thermocoupleInverseCoefficientsTypeS}
\alias{thermocoupleInverseCoefficientsTypeT}
\title{Polynomial Equation Coefficients for Voltage to Temperature for Thermocouple Type B}
\usage{thermocoupleInverseCoefficientsTypeB}
\description{\code{thermocoupleInverseCoefficientsTypeB} Coefficients for Voltage to Temperature for Thermocouple Type B
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(thermocoupleInverseCoefficientsTypeB)
str(thermocoupleInverseCoefficientsTypeB)
}
\keyword{datasets}
