\name{ThermocoupleInverseEquationTypeB}
\alias{ThermocoupleInverseEquationTypeB}
\alias{ThermocoupleInverseEquationTypeE}
\alias{ThermocoupleInverseEquationTypeJ}
\alias{ThermocoupleInverseEquationTypeK}
\alias{ThermocoupleInverseEquationTypeN}
\alias{ThermocoupleInverseEquationTypeR}
\alias{ThermocoupleInverseEquationTypeS}
\alias{ThermocoupleInverseEquationTypeT}
\title{Equation for Calculating Temperature from Voltage for Thermocouples Type B}
\usage{ThermocoupleInverseEquationTypeB(vV)}
\description{\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type B

\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type E

\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type J

\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type K

\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type N

\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type R

\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type S

\code{ThermocoupleInverseEquationTypeB} Calculates Voltage from Temperature for Thermocouples Type T
}
\arguments{
  \item{vV}{ Vector with voltages (C)}
}
\value{
  \item{T}{Temperature (C)}
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\keyword{programming}
