\name{thermocoupleInverseFunctionsRange}
\alias{thermocoupleInverseFunctionsRange}
\title{Error range for Polynomial inverse functions for Thermocouples}
\usage{thermocoupleInverseFunctionsRange}
\description{\code{thermocoupleInverseFunctionsRange} Error range for Polynomial inverse functions for Thermocouples
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(thermocoupleInverseFunctionsRange)
str(thermocoupleInverseFunctionsRange)
}
\keyword{datasets}
