\name{ thermocoupleMineralInsulated}
\alias{ thermocoupleMineralInsulated}
\title{Mineral Insulated Thermocouples}
\usage{ thermocoupleMineralInsulated}
\description{\code{ thermocoupleMineralInsulated} is a table with Mineral Insulated Thermocouples}
\source{
Watlow(R), 2014
Mineral Insulated Sensors by Diameter and Sheath
\url{https://www.watlow.com/downloads/en/catalogs/thermocouples.pdf}
}
\references{
Watlow(R), 2014
Mineral Insulated Sensors by Diameter and Sheath
\url{https://www.watlow.com/downloads/en/catalogs/thermocouples.pdf}
}
\author{Jose Gama}
\keyword{datasets}
