\name{ thermocoupleMounting}
\alias{ thermocoupleMounting}
\title{Thermocouple Mounting}
\usage{ thermocoupleMounting}
\description{\code{ thermocoupleMounting} is a table with Thermocouple Mounting data}
\source{
Capgo Pty Ltd, 2013
Thermocouple mounting
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
Watlow(R), 2014
Junction Types
\url{https://www.watlow.com/downloads/en/catalogs/thermocouples.pdf}
}
\references{
Capgo Pty Ltd, 2013
Thermocouple mounting
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
Watlow(R), 2014
Junction Types
\url{https://www.watlow.com/downloads/en/catalogs/thermocouples.pdf}
}
\author{Jose Gama}
\keyword{datasets}
