\name{thermocoupleNominalSeebeckCoefficients}
\alias{thermocoupleNominalSeebeckCoefficients}
\title{Nominal Seebeck Coefficients}
\usage{thermocoupleNominalSeebeckCoefficients}
\description{\code{thermocoupleNominalSeebeckCoefficients} is a table with Nominal Seebeck Coefficients
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(thermocoupleNominalSeebeckCoefficients)
str(thermocoupleNominalSeebeckCoefficients)
}
\keyword{datasets}
