\name{thermocoupleRecommendedUpperTempLimitsProtected}
\alias{thermocoupleRecommendedUpperTempLimitsProtected}
\title{Recommended Upper Temperature Limits for Protected Thermocouples}
\usage{thermocoupleRecommendedUpperTempLimitsProtected}
\description{\code{thermocoupleRecommendedUpperTempLimitsProtected} is a table with 
Recommended Upper Temperature Limits for Protected Thermocouples
}
\source{
CapGo, 2013
Recommended  upper temperature limits 
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\references{
CapGo, 2013
Recommended  upper temperature limits
\url{http://www.capgo.com/Resources/Temperature/Thermocouple/Thermocouple.html}
}
\author{Jose Gama}
\examples{
data(thermocoupleRecommendedUpperTempLimitsProtected)
str(thermocoupleRecommendedUpperTempLimitsProtected)
}
\keyword{datasets}
