\name{ thermocoupleResponseTime}
\alias{ thermocoupleResponseTime}
\title{Thermocouple Response Times}
\usage{ thermocoupleResponseTime}
\description{\code{ thermocoupleResponseTime} is a table with Thermocouple Response Times}
\source{
Industrial Temperature Sensors Ltd., 2014
Typical Thermocouple Response Times in seconds
\url{http://www.itsirl.com/tcresp.php}
}
\references{
Industrial Temperature Sensors Ltd., 2014
Typical Thermocouple Response Times in seconds
\url{http://www.itsirl.com/tcresp.php}
}
\author{Jose Gama}
\keyword{datasets}
