\name{thermocoupleSingleLegThermoelements}
\alias{thermocoupleSingleLegThermoelements}
\title{Letter designations, compositions, and trade names of single-leg thermoelements}
\usage{thermocoupleSingleLegThermoelements}
\description{\code{thermocoupleSingleLegThermoelements} is a table with the Letter designations, 
compositions, and trade names of single-leg thermoelements}
\source{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_ii.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
Table I Thermocouple Types Definitions
\url{http://srdata.nist.gov/its90/tables/table_ii.html}
}
\author{Jose Gama}
\examples{
data(thermocoupleSingleLegThermoelements)
str(thermocoupleSingleLegThermoelements)
}
\keyword{datasets}
