\name{ ThermocoupleTable10colsTo2}
\alias{ ThermocoupleTable10colsTo2}
\title{Convert the thermocouple table for easier use}
\usage{ ThermocoupleTable10colsTo2(thermocoupleTable)}
\description{\code{ ThermocoupleTable10colsTo2} converts the thermocouple table from n X 12 to m X 2}
\arguments{
  \item{thermocoupleTable}{thermocouple table to be resized n X 12}
}
\value{
  \item{table}{thermocouple table m X 2}
}
\author{Jose Gama}
\keyword{programming}
