\name{thermocoupleTypeBthermoelectricVoltage}
\alias{thermocoupleTypeBthermoelectricVoltage}
\alias{thermocoupleTypeEthermoelectricVoltage}
\alias{thermocoupleTypeJthermoelectricVoltage}
\alias{thermocoupleTypeKthermoelectricVoltage}
\alias{thermocoupleTypeNthermoelectricVoltage}
\alias{thermocoupleTypeRthermoelectricVoltage}
\alias{thermocoupleTypeSthermoelectricVoltage}
\alias{thermocoupleTypeTthermoelectricVoltage}
\title{Thermoelectric Voltage for Thermocouple Type B}
\usage{thermocoupleTypeBthermoelectricVoltage}
\description{\code{thermocoupleTypeBthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type B

\code{thermocoupleTypeEthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type E

\code{thermocoupleTypeJthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type J

\code{thermocoupleTypeKthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type K

\code{thermocoupleTypeNthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type N

\code{thermocoupleTypeSthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type R

\code{thermocoupleTypeBthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type S

\code{thermocoupleTypeTthermoelectricVoltage} Thermoelectric Voltage for Thermocouple Type T
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(thermocoupleTypeBthermoelectricVoltage)
str(thermocoupleTypeBthermoelectricVoltage)
}
\keyword{datasets}
