\name{thermocoupleTypesASTM}
\alias{thermocoupleTypesASTM}
\title{Thermocouple Wire Constituents}
\usage{thermocoupleTypesASTM}
\description{\code{thermocoupleTypesASTM} is a table with Thermocouple Wire Constituents according to the ASTM
}
\source{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\references{
Kerlin, T.W., 1999
Practical Thermocouple Thermometry
International Society of Automation (ISA)
}
\author{Jose Gama}
\examples{
data(thermocoupleTypesASTM)
str(thermocoupleTypesASTM)
}
\keyword{datasets}
