\name{thermocoupleWireColorUnitedStatesCanadaMexico}
\alias{thermocoupleWireColorUnitedStatesCanadaMexico}
\alias{thermocoupleCompensatingExtensionWireColorUnitedStatesCanadaMexico}
\alias{thermocoupleExtensionWireColorUnitedStatesCanadaMexico}
\alias{thermocoupleAndExtensionWiresInternationalColorCodes}
\title{Wire Color for Thermocouples and Lead Wires}
\usage{thermocoupleWireColorUnitedStatesCanadaMexico}
\description{\code{thermocoupleWireColorUnitedStatesCanadaMexico} Wire Color for 
Thermocouples and Lead Wires for the United States, Canada and Mexico

\code{thermocoupleCompensatingExtensionWireColorUnitedStatesCanadaMexico} Wire Color for 
Compensating Extension Wire for the United States, Canada and Mexico

\code{thermocoupleExtensionWireColorUnitedStatesCanadaMexico} Wire Color for 
Extension Wire for the United States, Canada and Mexico

\code{thermocoupleAndExtensionWiresInternationalColorCodes} Wire Color for 
Thermocouples and Extension Wires with international codes

}
\source{
TEMPCO Electric Heater Corporation, 2014
Temperature Sensing
\url{www.tempco.com}

American Society for Testing and Materials, 1993
Manual on the Use of Thermocouples in Temperature Measurement
Committee E20 on Temperature Measurement and Subcommittee E20.04 on Thermocouples
}
\references{
TEMPCO Electric Heater Corporation, 2014
Temperature Sensing
\url{www.tempco.com}

American Society for Testing and Materials, 1993
Manual on the Use of Thermocouples in Temperature Measurement
Committee E20 on Temperature Measurement and Subcommittee E20.04 on Thermocouples
}
\author{Jose Gama}
\examples{
data(thermocoupleWireColorUnitedStatesCanadaMexico)
str(thermocoupleWireColorUnitedStatesCanadaMexico)
}
\keyword{datasets}
