\name{thermocoupleWireSizeResistanceImperial}
\alias{thermocoupleWireSizeResistanceImperial}
\title{thermocouple wire size and resistance table}
\usage{thermocoupleWireSizeResistanceImperial}
\description{\code{thermocoupleWireSizeResistanceImperial} is a table with thermocouple wire size and resistance}
\source{
Conax(TM) Buffalo, 2014
thermocouple wire size and resistance table
\url{www.conaxbuffalo.com}
}
\references{
Conax(TM) Buffalo, 2014
thermocouple wire size and resistance table
\url{www.conaxbuffalo.com}
}
\author{Jose Gama}
\examples{
data(thermocoupleWireSizeResistanceImperial)
str(thermocoupleWireSizeResistanceImperial)
}
\keyword{datasets}
