\name{ TminusT90CCT2008}
\alias{ TminusT90CCT2008}
\title{T - T90 computed by a polynomial}
\usage{ TminusT90CCT2008(T90K)}
\description{\code{ TminusT90CCT2008} Thermodynamic Temperature minus the ITS-90,
 computed by a polynomial (CCT WG4 2008)}
\arguments{
  \item{T90K}{ ITS-90}
}
\value{
  \item{T - T90}{ Thermodynamic Temperature minus ITS-90}
}
\source{
Franco Pavese and Gianfranco Molinar Min Beciet, 2013
Modern Gas-Based Temperature and Pressure Measurements
Springer Science + Business Media
pp. 42
}
\references{
Franco Pavese and Gianfranco Molinar Min Beciet, 2013
Modern Gas-Based Temperature and Pressure Measurements
Springer Science + Business Media
pp. 42
}
\author{Jose Gama}
\keyword{programming}
