\name{ TminusT90Pavese6CubicPolynomials}
\alias{ TminusT90Pavese6CubicPolynomials}
\title{T - T90 computed by 6 cubic polynomials}
\usage{ TminusT90Pavese6CubicPolynomials(T90K)}
\description{\code{ TminusT90Pavese6CubicPolynomials} Thermodynamic Temperature minus 
the ITS-90, computed by 6 cubic polynomials (CCT WG4 2008)}
\arguments{
  \item{T90K}{ ITS-90}
}
\value{
  \item{T - T90}{ Thermodynamic Temperature minus ITS-90}
}
\source{
Franco Pavese and Gianfranco Molinar Min Beciet, 2013
Modern Gas-Based Temperature and Pressure Measurements
Springer Science + Business Media
pp. 42
}
\references{
Franco Pavese and Gianfranco Molinar Min Beciet, 2013
Modern Gas-Based Temperature and Pressure Measurements
Springer Science + Business Media
pp. 42
}
\author{Jose Gama}
\keyword{programming}
